Improve perfomance of height updates using rxjs
When the grid size is set to a larger number it can happen that
have to render round about four thousands `HexagonPillar` components. If
each of them need to be rerendered by react, if we change the sea level or
height factor, it gets very laggy.

Now we use a central instance of a RxJS subject which is updated by recoil using
atom effects. These subjects can be subscribed by all `HexagonPillars` and update
the threejs instances directly. So, we don't need a react rerender in this cases
and everything works on 60fps.

Maybe react concurrent mode will solve this in the future. Maybe we will not have
such problems when we begin to render the pillars using `InstancedMesh` later.